Ended (probably) refactoring
