+++ SignalR_UnitTestingSupportCommon/Hubs/HubUnitTestsSupportT.cs
@@ -1,13 +1,13 @@
 ﻿using Microsoft.AspNetCore.SignalR;
 using Moq;
 using SignalR_UnitTestingSupportCommon.Hubs.Internal;
+using SignalR_UnitTestingSupportCommon.Interfaces;
 using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace SignalR_UnitTestingSupportCommon.Hubs
 {
-    public class HubUnitTestsSupport<TIHubResponses> : HubUnitTestsBaseCommon
+    public abstract class HubUnitTestsSupport<TIHubResponses> : HubUnitTestsBaseCommon, IHubUnitTestsBase<TIHubResponses>
         where TIHubResponses : class
     {
         public Mock<IHubCallerClients<TIHubResponses>> ClientsMock { get; private set; }
@@ -25,29 +25,12 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
         public Mock<TIHubResponses> ClientsUserMock { get; private set; }
         public Mock<TIHubResponses> ClientsUsersMock { get; private set; }
 
-        internal override void _setUpClients()
-        {
-            SetUpClients();
-            SetUpClientsAll();
-            SetUpClientsAllExcept();
-            SetUpClientsCaller();
-            SetUpClientsClient();
-            SetUpClientsClients();
-            SetUpClientsGroup();
-            SetUpClientsGroupExcept();
-            SetUpClientsGroups();
-            SetUpClientsOthersMock();
-            SetUpClientsOthersInGroup();
-            SetUpClientsUser();
-            SetUpClientsUsers();
-        }
-
-        internal void SetUpClients()
+        internal override void SetUpClients()
         {
             ClientsMock = new Mock<IHubCallerClients<TIHubResponses>>();
         }
 
-        internal  void SetUpClientsUsers()
+        internal override void SetUpClientsUsers()
         {
             ClientsUsersMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -55,7 +38,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsUsersMock.Object);
         }
 
-        internal  void SetUpClientsUser()
+        internal override void SetUpClientsUser()
         {
             ClientsUserMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -63,7 +46,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsUserMock.Object);
         }
 
-        internal  void SetUpClientsOthersInGroup()
+        internal override void SetUpClientsOthersInGroup()
         {
             ClientsOthersInGroupMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -71,7 +54,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsOthersInGroupMock.Object);
         }
 
-        internal  void SetUpClientsOthersMock()
+        internal override void SetUpClientsOthersMock()
         {
             ClientsOthersMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -79,7 +62,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsOthersMock.Object);
         }
 
-        internal  void SetUpClientsGroups()
+        internal override void SetUpClientsGroups()
         {
             ClientsGroupsMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -87,7 +70,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsGroupsMock.Object);
         }
 
-        internal  void SetUpClientsGroupExcept()
+        internal override void SetUpClientsGroupExcept()
         {
             ClientsGroupExceptMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -95,7 +78,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsGroupExceptMock.Object);
         }
 
-        internal  void SetUpClientsGroup()
+        internal override void SetUpClientsGroup()
         {
             ClientsGroupMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -103,7 +86,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsGroupMock.Object);
         }
 
-        internal  void SetUpClientsClients()
+        internal override void SetUpClientsClients()
         {
             ClientsClientsMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -111,7 +94,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsClientsMock.Object);
         }
 
-        internal  void SetUpClientsClient()
+        internal override void SetUpClientsClient()
         {
             ClientsClientMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -119,7 +102,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsClientMock.Object);
         }
 
-        internal  void SetUpClientsCaller()
+        internal override void SetUpClientsCaller()
         {
             ClientsCallerMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -127,7 +110,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsCallerMock.Object);
         }
 
-        internal  void SetUpClientsAllExcept()
+        internal override void SetUpClientsAllExcept()
         {
             ClientsAllExceptMock = new Mock<TIHubResponses>();
             ClientsMock
@@ -135,7 +118,7 @@ namespace SignalR_UnitTestingSupportCommon.Hubs
                 .Returns(ClientsAllExceptMock.Object);
         }
 
-        internal  void SetUpClientsAll()
+        internal override void SetUpClientsAll()
         {
             ClientsAllMock = new Mock<TIHubResponses>();
             ClientsMock
