+++ SignalR_UnitTestingSupportMSTest/SignalR_UnitTestingSupportMSTest.csproj
@@ -53,4 +53,8 @@
     <PackageReference Include="MSTest.TestFramework" Version="1.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\SignalR_UnitTestingSupportCommon\SignalR_UnitTestingSupportCommon.csproj" />
+  </ItemGroup>
+
 </Project>
