+++ SignalR_UnitTestingSupportXUnit/Hubs/HubUnitTestsWithEF`.cs
@@ -1,8 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Moq;
-using Xunit;
-using SignalR_UnitTestingSupportXUnit.Services;
-
+using SignalR_UnitTestingSupportCommon.Services;
 
 namespace SignalR_UnitTestingSupportXUnit.Hubs
 {
