+++ SignalR_UnitTestingSupportXUnit/SignalR_UnitTestingSupportXUnit.csproj
@@ -53,4 +53,8 @@
     <PackageReference Include="xunit" Version="2.4.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\SignalR_UnitTestingSupportCommon\SignalR_UnitTestingSupportCommon.csproj" />
+  </ItemGroup>
+
 </Project>
